feat: update protocol 15.24 (#129)
* update

* feat(protocol): update protocol from 14.12 to 15.11

- Update client version to 15.11 (1511)
- Rewrite onConnect challenge packet format (U8 length, 0x71 extra byte)
- Rewrite onRecvFirstMessage for 15.11 (asset hash, OTCv8 detection)
- Complete AddPlayerSkills rewrite (percent*100, combat stats as doubles)
- Reorder login packet sequence to match 15.11 protocol
- Add new stub packets (bosstiary, prey, forge, proficiency, VIP groups, etc.)
- Update MagicEffectClasses/ShootType_t from uint8_t to uint16_t throughout
- Update wire format to use uint16_t for magic/distance effects in loop system
- Add new enums: InspectObjectType, Imbuement_Window_t, WeaponProficiency_t
- Expand NETWORKMESSAGE_MAXSIZE to 65500, Icons_t, ResourceTypes_t


* fix(protocol): correct 5 byte-level issues in 15.11 login packets

- Fix challenge packet (0x1F): skipBytes(-13) -> (-12), checksum 9 -> 8
- Fix sendIcons (0xA2): icons field from U32 to U64 (client expects 8 bytes)
- Fix AddPlayerSkills (0xA1): addDouble precision from 2/3 to 4
- Fix sendWorldLight (0x82): add isAccessPlayer check for GM brightness
- Fix sendPremiumTrigger (0x9E): send 16 triggers instead of empty list


* update

* Update protocolgame.cpp

* update

* Update protocolgame.cpp

* update

* Update outputmessage.h

* refactor(bestiary): simplify bestiary message, add fallback for missing param

* chore(resources): add new resource constants to cpplinter

* chore(protocolgame): add TODO for needed bytes in empty containers

* cleanup(protocol): remove checksum mode handling, force sequence mode

* refactor(protocol): remove unused sendTibiaTime and sendWorldLight methods

* consistency fixes

* use uint64_t for all Icon_* bitmasks and icon retrieval paths

The Icons_t enum used the default int underlying type while Player::getClientIcons
and other consumers accumulated icons into a uint64_t. This created inconsistent
mixing of int and uint64_t types across the codebase.

- Set enum Icons_t : uint64_t in src/const.h so all ICON_* constants are uint64_t
- Change Condition::getIcons() (base + 5 overrides) to return uint64_t
- Update all local icon variables in condition.cpp to uint64_t
- Fix ternary literal type (0 → uint64_t{0}) to satisfy -Wextra

* fix: mismatched enum type in Condition::getIcons

---------

+++ src/combat.h
@@ -53,8 +53,8 @@ struct CombatParams
 	CombatType_t combatType = COMBAT_NONE;
 	CombatOrigin origin = ORIGIN_SPELL;
 
-	uint8_t impactEffect = CONST_ME_NONE;
-	uint8_t distanceEffect = CONST_ANI_NONE;
+	uint16_t impactEffect = CONST_ME_NONE;
+	uint16_t distanceEffect = CONST_ANI_NONE;
 
 	bool blockedByArmor = false;
 	bool blockedByShield = false;
@@ -100,7 +100,7 @@ public:
 	                              const CombatParams& params);
 
 	static void addDistanceEffect(const std::shared_ptr<Creature>& caster, const Position& fromPos,
-	                              const Position& toPos, uint8_t effect);
+	                              const Position& toPos, uint16_t effect);
 
 	void doCombat(const std::shared_ptr<Creature>& caster, const std::shared_ptr<Creature>& target) const;
 	void doCombat(const std::shared_ptr<Creature>& caster, const Position& position) const;
